# docs/filesir-swagger.yaml
openapi: 3.0.0
info:
  description: 'Access token should be sent along with every request to Files.ir API in authorization header: `Authorization: Bearer <Token>` <br> This token can either be acquired via `auth/login` endpoint or by going to account settings > developers on [Files.ir website](SITE_URL/account-settings/#developers).'
  version: '1.0.0'
  title: Files.ir API
security:
  - accessToken: []
...
paths:
...
  /uploads:
    post:
      tags:
        - Uploads
      summary: Upload a file to Files.ir
      operationId: upload
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Content of file to upload to Files.ir
                parentId:
                  type: integer
                  description: ID of folder where this file should be uploaded, `null` will upload to root
                  example: null
                workspaceId:
                  type: integer
                  description: ID of workspace where this file should be uploaded. If omitted or `0`, upload goes to personal drive.
                  example: 123
                relativePath:
                  type: string
                  description: "Folders in the path provided here will be auto created, if they don't exist already. This is mainly useful when uploading a folder from browser. It should include original filename as well: <br> `/some/folders/here/file-name.jpg`"
      responses:
        '201':
          description: File was uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  fileEntry:
                    $ref: '#/components/schemas/FileEntry'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
        '422':
          description: Invalid data specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/422-Response'
...
components:
  schemas:
...
    FileEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: image.png
        file_name:
          type: string
          example: j6oWHDG1z613UqFHLZpE1Yrqngw0N8XLuGNLNWJV
          description: actual file name for this entry on storage
        file_size:
          type: integer
          format: int64
          example: 111863
          description: In bytes
        parent_id:
          type: integer
          format: int64
          example: 1
          description: Parent folder ID
        workspace_id:
          type: integer
          format: int64
          example: 0
          description: Workspace ID this entry belongs to. `0` means personal drive.
        parent:
          $ref: '#/components/schemas/FileEntry'
        thumbnail:
          type: string
          example: thumbnail.png
          description: Relative path to thumbnail image for the file (if it exists)
        mime:
          type: string
          example: image/png
        url:
          type: string
          example: secure/uploads/3260
          description: relative url for previewing file or folder contents
        hash:
          type: string
          example: MzI2MHxwYWRkaQ
        type:
          type: string
          enum:
            - 'image'
            - 'folder'
            - 'text'
            - 'audio'
            - 'video'
            - 'pdf'
        description:
          type: string
          example: Example description entered by user
        deleted_at:
          type: string
          example: '2021-02-23T14:42:38.000000Z'
          description: date entry was moved to trash (if it's in trash currently)
        created_at:
          type: string
          example: '2021-02-23T14:42:38.000000Z'
          description: date entry was uploaded
        updated_at:
          type: string
          example: '2021-02-23T14:42:38.000000Z'
          description: date entry was last modified (content, name or location changed)
        path:
          type: string
          example: '3260/3261/3262'
          description: full path of parent folder IDs for this entry up to root
        users:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              email:
                type: string
                example: 'example@gmail.com'
...
    401-Response:
      description: Unauthenticated. Either access token is not provided or is invalid.
    403-Response:
      description: Unauthorized access. You don't have permissions required to perform this operation.
    422-Response:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: 'Reason for the error'
        errors:
          type: object
          properties:
            some_data_1:
              type: string
              example: Error message for data 1
            some_data_2:
              type: string
              example: Error message for data 2
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer